Got the CPP version generating combinations.
So far it's only combinations of non-negative integers less than some maximum, but that's all we need because they're going to be array (or vector) indexes...

The other limitation is that it calculates them sequentially. There's no way yet to calculate the Nth combination directly, which we WILL need for splitting the work up across threads.