Rewrote the high hand scoring function to make it much faster.
It now represents the hand as a bit-field for each suit. This lets us calculate the set of distinct cards, and refine that to find out the exact hand and calculate a score, using simple bitwise operations. No looping required, no multiplying (but a bit of bit-shifting).

For the 45-card case it now runs in approx 130 seconds, down from 356 seconds previously.